+++ example/lib/provider/linkfive_provider.dart
@@ -2,10 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:in_app_purchases_paywall_ui/in_app_purchases_paywall_ui.dart';
-import 'package:linkfive_purchases/logic/linkfive_purchases_main.dart';
-import 'package:linkfive_purchases/models/linkfive_active_subscription.dart';
-import 'package:linkfive_purchases/models/linkfive_response.dart';
-import 'package:linkfive_purchases/models/linkfive_subscription.dart';
+import 'package:linkfive_purchases/purchases.dart';
 import 'package:linkfive_purchases_example/key/keyLoader.dart';
 
 class LinkFiveProvider extends ChangeNotifier {
@@ -18,11 +15,17 @@ class LinkFiveProvider extends ChangeNotifier {
   List<StreamSubscription> _streams = [];
 
   LinkFiveProvider(Keys keys) {
-    linkFivePurchases.init(keys.linkFiveApiKey, env: LinkFiveEnvironment.STAGING);
+    linkFivePurchases.init(keys.linkFiveApiKey,
+        env: LinkFiveEnvironment.STAGING);
     linkFivePurchases.fetchSubscriptions();
-    _streams.add(linkFivePurchases.listenOnResponseData().listen(_responseDataUpdate));
-    _streams.add(linkFivePurchases.listenOnSubscriptionData().listen(_subscriptionDataUpdate));
-    _streams.add(linkFivePurchases.listenOnActiveSubscriptionData().listen(_activeSubscriptionDataUpdate));
+    _streams.add(
+        linkFivePurchases.listenOnResponseData().listen(_responseDataUpdate));
+    _streams.add(linkFivePurchases
+        .listenOnSubscriptionData()
+        .listen(_subscriptionDataUpdate));
+    _streams.add(linkFivePurchases
+        .listenOnActiveSubscriptionData()
+        .listen(_activeSubscriptionDataUpdate));
   }
 
   void _responseDataUpdate(LinkFiveResponseData? data) {