+++ example/lib/routing/delegate.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:linkfive_purchases/linkfive_purchases.dart';
 import 'package:linkfive_purchases_example/page/raw_page.dart';
 import 'package:linkfive_purchases_example/page/simple_paywall_ui_page.dart';
 import 'package:linkfive_purchases_example/root_page.dart';
 import 'package:linkfive_purchases_example/routing/app_path.dart';
 import 'package:linkfive_purchases_example/routing/no_animation_transition_delegate.dart';
 
-class MainRouterDelegate extends RouterDelegate<AppPath> with ChangeNotifier, PopNavigatorRouterDelegateMixin<AppPath> {
+class MainRouterDelegate extends RouterDelegate<AppPath>
+    with ChangeNotifier, PopNavigatorRouterDelegateMixin<AppPath> {
   final GlobalKey<NavigatorState> navigatorKey;
 
   bool isRawPaywall = false;
@@ -16,7 +16,8 @@ class MainRouterDelegate extends RouterDelegate<AppPath> with ChangeNotifier, Po
 
   MainRouterDelegate() : navigatorKey = GlobalKey<NavigatorState>();
 
-  AppPath get currentConfiguration => isSimplePaywall ? AppPath.home() : AppPath.simplePaywall();
+  AppPath get currentConfiguration =>
+      isSimplePaywall ? AppPath.home() : AppPath.simplePaywall();
 
   @override
   Widget build(BuildContext context) {
@@ -26,8 +27,7 @@ class MainRouterDelegate extends RouterDelegate<AppPath> with ChangeNotifier, Po
       pages: [
         RootPage(),
         if (isRawPaywall) RawPage(),
-        if (isSimplePaywall) SimplePayWallUiPage(),
-        //if (_selectedRecipeId != null) RecipeDetailPage(_selectedRecipeId!)
+        if (isSimplePaywall) SimplePaywallUiPage()
       ],
       onPopPage: (route, result) {
         if (!route.didPop(result)) {
@@ -37,7 +37,7 @@ class MainRouterDelegate extends RouterDelegate<AppPath> with ChangeNotifier, Po
         if (isSimplePaywall) {
           isSimplePaywall = false;
         }
-        if(isRawPaywall){
+        if (isRawPaywall) {
           isRawPaywall = false;
         }
         notifyListeners();
@@ -64,7 +64,7 @@ class MainRouterDelegate extends RouterDelegate<AppPath> with ChangeNotifier, Po
 
   goToSimplePayWall() {
     isSimplePaywall = true;
-    LinkFivePurchases.fetchSubscriptions();
+    // LinkFivePurchases.fetchSubscriptions();
     notifyListeners();
   }
 }