+++ lib/models/linkfive_active_subscription.dart
@@ -1,6 +1,8 @@
 import 'package:linkfive_purchases/models/linkfive_verified_receipt.dart';
 
+/// Model for subscriptions to offer to the user
 class LinkFiveActiveSubscriptionData {
+  /// these subscriptions are available for purchase
   final List<LinkFiveVerifiedReceipt> subscriptionList;
 
   LinkFiveActiveSubscriptionData(this.subscriptionList);