+++ lib/models/linkfive_response.dart
@@ -1,3 +1,4 @@
+/// Response Data from LinkFive
 class LinkFiveResponseData {
   final String platform;
   final String? attributes;
@@ -8,8 +9,9 @@ class LinkFiveResponseData {
   LinkFiveResponseData.fromJson(Map<Object?, Object?> json)
       : platform = json["platform"] as String,
         attributes = json["attributes"] as String?,
-        subscriptionList = (json["subscriptionList"] as List).map((e) => LinkFiveResponseDataSubscription.fromJson(e)).toList()
-  ;
+        subscriptionList = (json["subscriptionList"] as List)
+            .map((e) => LinkFiveResponseDataSubscription.fromJson(e))
+            .toList();
 }
 
 class LinkFiveResponseDataSubscription {
@@ -17,5 +19,6 @@ class LinkFiveResponseDataSubscription {
 
   LinkFiveResponseDataSubscription(this.sku);
 
-  LinkFiveResponseDataSubscription.fromJson(Map<Object?, Object?> json) : sku = json["sku"] as String;
+  LinkFiveResponseDataSubscription.fromJson(Map<Object?, Object?> json)
+      : sku = json["sku"] as String;
 }