+++ lib/purchases.dart
@@ -1,9 +1,10 @@
 library linkfive_purchases;
 
 export 'linkfive_purchases.dart';
+export 'logic/linkfive_purchases_main.dart';
 export 'models/linkfive_active_subscription.dart';
 export 'models/linkfive_response.dart';
 export 'models/linkfive_subscription.dart';
 export 'models/linkfive_verified_receipt.dart';
-export 'models/subscription_period.dart';
+export 'util/subscription_duration_convert.dart';
 export 'package:in_app_purchases_interface/in_app_purchases_interface.dart';