+++ lib/store/linkfive_app_data_store.dart
@@ -1,3 +1,4 @@
+/// Model to store some sdk data
 class LinkFiveAppDataStore {
   String apiKey = "";
   String? utmSource = null;