+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+add easy paywall ui integration.
+add provider example.
+add streambuilder example.
+add intl package.
+add interface package
+
 ## 1.0.1
 * fixed a bug where a purchase was not immediately processed as a purchase.